+++ packages/components/drawer/src/__specs__/Drawer.spec.tsx
@@ -1,26 +1,23 @@
 import * as React from 'react';
 import { renderWithProvider } from '@synerise/ds-utils';
 import Typography from 'antd/lib/typography';
-import { DrawerBody, DrawerContent, DrawerHeader } from '../Drawer.styles';
 import Drawer from '../Drawer';
 
 const DRAWER = (visible: boolean) => (<Drawer
   visible={visible}
-  placement='right'
   width={400}
+  placement='right'
 >
-  <>
-    <DrawerHeader>
-      <div style={{display: 'flex', justifyContent: 'space-between', alignItems: 'center', paddingBottom: 24}}>
-        <Typography.Title style={{flex: 1, margin: 0,}} level={4}>Example drawer</Typography.Title>
-      </div>
-    </DrawerHeader>
-    <DrawerBody>
-      <DrawerContent>
-        <p>Content</p>
-      </DrawerContent>
-    </DrawerBody>
-  </>
+  <Drawer.DrawerHeader>
+    <div style={{display: 'flex', justifyContent: 'space-between', alignItems: 'center', paddingBottom: 24}}>
+      <Typography.Title style={{flex: 1, margin: 0,}} level={4}>Example drawer</Typography.Title>
+    </div>
+  </Drawer.DrawerHeader>
+  <Drawer.DrawerBody>
+    <Drawer.DrawerContent>
+      <p>Content</p>
+    </Drawer.DrawerContent>
+  </Drawer.DrawerBody>
 </Drawer>);
 
 describe('Drawer component', () => {