+++ packages/portal/stories/components/ManageableList/index.stories.tsx
@@ -4,7 +4,6 @@ import FileM from '@synerise/ds-icon/dist/icons/FileM';
 import Tag, { TagShape } from '@synerise/ds-tags/dist/Tag/Tag';
 import { withState } from '@dump247/storybook-state';
 import { action } from '@storybook/addon-actions';
-import { ListType } from '@synerise/ds-manageable-list/dist/ManageableList';
 import { boolean } from '@storybook/addon-knobs';
 import FolderM from '@synerise/ds-icon/dist/icons/FolderM';
 
@@ -15,6 +14,23 @@ const decorator = (storyFn) => (
     </div>
   </div>
 );
+
+const removeItem = (props, store): void => {
+  store.set({
+    items: store.state.items.filter(item => item.id !== props.id),
+  });
+};
+
+const editItem = (props, store): void => {
+  store.set({
+    items: store.state.items.map(item => {
+      if(item.id === props.id) {
+        item.name = props.name;
+      }
+      return item;
+    })
+  })
+};
 
 const ITEMS:any = [
   {
@@ -114,6 +130,16 @@ const CONTENT_ITEMS: any = [
   }
 ];
 
+const EMPTY_CONTENT_ITEM = {
+    id: "",
+    name: "New Item",
+    canAdd: true,
+    canUpdate: true,
+    canDelete: true,
+    tag: <Tag name={"A"} shape={TagShape.SINGLE_CHARACTER_ROUND} color={"red"} />,
+    content: <div>content</div>
+};
+
 const stories = {
   default: withState({
     items: ITEMS,
@@ -131,23 +157,6 @@ const stories = {
       });
     };
 
-    const removeItem = ({id}): void => {
-      store.set({
-        items: store.state.items.filter(item => item.id !== id),
-      });
-    };
-
-    const editItem = (props): void => {
-      store.set({
-        items: store.state.items.map(item => {
-          if(item.id === props.id) {
-            item.name = props.name;
-          }
-          return item;
-        })
-      })
-    };
-
     return (
       <ManageableList
         addItemLabel="Add folder"
@@ -157,8 +166,8 @@ const stories = {
         less="less"
         maxToShowItems={5}
         onItemAdd={addItem}
-        onItemRemove={removeItem}
-        onItemEdit={editItem}
+        onItemRemove={props => removeItem(props, store)}
+        onItemEdit={props => editItem(props, store)}
         onItemSelect={action('onItemSelect')}
         items={store.state.items}
         loading={false}
@@ -186,6 +195,31 @@ const stories = {
       store.set({items: newOrder});
     };
 
+    const addItem = (): void => {
+      store.set({
+        items: [
+          ...store.state.items,
+          {
+            ...EMPTY_CONTENT_ITEM,
+            id: Date.now(),
+          }
+        ]
+      });
+    };
+
+    const duplicateItem = (props): void => {
+      const itemForDuplication = store.state.items.find(item => item.id === props.id);
+      store.set({
+        items: [
+          ...store.state.items,
+          {
+            ...itemForDuplication,
+            id: Date.now(),
+          }
+        ]
+      })
+    };
+
     return (
       <ManageableList
         addItemLabel="Add position"
@@ -194,13 +228,13 @@ const stories = {
         more="more"
         less="less"
         maxToShowItems={5}
-        onItemAdd={action('onItemAdd')}
-        onItemRemove={action('onItemRemove')}
-        onItemEdit={action('onItemEdit')}
+        onItemAdd={addItem}
+        onItemRemove={props => removeItem(props, store)}
+        onItemEdit={props => editItem(props, store)}
         onItemSelect={action('onItemSelect')}
-        onItemDuplicate={action('onItemDuplicate')}
+        onItemDuplicate={duplicateItem}
         onChangeOrder={boolean('Change order available', false) ? handleChangeOrder : null}
-        type={ListType.content}
+        type='content'
         items={store.state.items}
         loading={false}
         addButtonDisabled={boolean('Disable add item button', false)}