+++ scripts/docs/componentmd.js
@@ -10,18 +10,26 @@ const DOCS = __dirname + '/../../packages/docs-site/docs/components';
 const newSidebars = sidebars;
 const omit = ['core', 'utils'];
 
+const copyFileToDocs = (file, cb) => {
+  const dir = path
+    .dirname(file)
+    .split(path.sep)
+    .pop();
+  if (!omit.includes(dir)) {
+    const destinationFile = path.resolve(DOCS + `/${dir}.md`);
+    fs.createFileSync(destinationFile);
+    fs.copySync(file, destinationFile);
+    cb && cb(dir);
+  }
+};
+
+module.exports = copyFileToDocs;
+
 glob(COMPONENTS + '/*/README.md', { nodir: true }, function(err, files) {
   files.forEach(file => {
-    const dir = path
-      .dirname(file)
-      .split(path.sep)
-      .pop();
-    if (!omit.includes(dir)) {
-      const destinationFile = path.resolve(DOCS + `/${dir}.md`);
-      fs.createFileSync(destinationFile);
-      fs.copySync(file, destinationFile);
+    copyFileToDocs(file, dir => {
       newSidebars.docs.Components.push(`components/${dir}`);
-    }
+    });
   });
   fs.writeFileSync(path.resolve(DOCS, '../../website/sidebars.json'), JSON.stringify(newSidebars), 'utf-8');
 });