+++ packages/components/card-tabs/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.7](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card-tabs@0.1.6...@synerise/ds-card-tabs@0.1.7) (2019-11-15)
+
+**Note:** Version bump only for package @synerise/ds-card-tabs
+
+## [0.1.5](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card-tabs@0.1.4...@synerise/ds-card-tabs@0.1.5) (2019-11-12)
+
+**Note:** Version bump only for package @synerise/ds-card-tabs
+
+## [0.1.4](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card-tabs@0.1.3...@synerise/ds-card-tabs@0.1.4) (2019-11-09)
+
+**Note:** Version bump only for package @synerise/ds-card-tabs
+
 ## [0.1.3](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card-tabs@0.1.2...@synerise/ds-card-tabs@0.1.3) (2019-11-06)
 
 **Note:** Version bump only for package @synerise/ds-card-tabs