+++ packages/components/card/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.11](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card@0.2.10...@synerise/ds-card@0.2.11) (2019-11-15)
+
+**Note:** Version bump only for package @synerise/ds-card
+
+## [0.2.9](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card@0.2.8...@synerise/ds-card@0.2.9) (2019-11-09)
+
+**Note:** Version bump only for package @synerise/ds-card
+
 ## [0.2.8](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card@0.2.7...@synerise/ds-card@0.2.8) (2019-11-06)
 
 **Note:** Version bump only for package @synerise/ds-card