+++ packages/components/dropdown/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.15](https://github.com/synerise/synerise-design/compare/@synerise/ds-dropdown@0.7.14...@synerise/ds-dropdown@0.7.15) (2019-11-18)
+
+**Note:** Version bump only for package @synerise/ds-dropdown
+
+## [0.7.14](https://github.com/synerise/synerise-design/compare/@synerise/ds-dropdown@0.7.13...@synerise/ds-dropdown@0.7.14) (2019-11-15)
+
+**Note:** Version bump only for package @synerise/ds-dropdown
+
+## [0.7.12](https://github.com/synerise/synerise-design/compare/@synerise/ds-dropdown@0.7.11...@synerise/ds-dropdown@0.7.12) (2019-11-09)
+
+**Note:** Version bump only for package @synerise/ds-dropdown
+
 ## [0.7.11](https://github.com/synerise/synerise-design/compare/@synerise/ds-dropdown@0.7.10...@synerise/ds-dropdown@0.7.11) (2019-11-06)
 
 **Note:** Version bump only for package @synerise/ds-dropdown