+++ packages/components/navbar/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/Synerise/synerise-design/compare/@synerise/ds-navbar@0.1.2...@synerise/ds-navbar@0.1.3) (2019-11-15)
+
+**Note:** Version bump only for package @synerise/ds-navbar
+
+## [0.1.1](https://github.com/Synerise/synerise-design/compare/@synerise/ds-navbar@0.1.0...@synerise/ds-navbar@0.1.1) (2019-11-09)
+
+**Note:** Version bump only for package @synerise/ds-navbar
+
 # 0.1.0 (2019-11-06)
 
 ### Features