+++ packages/components/slider/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.7](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.1.6...@synerise/ds-slider@0.1.7) (2019-11-12)
+
+### Bug Fixes
+
+- **slider:** add styling fixes to slider ([3596448](https://github.com/Synerise/synerise-design/commit/35964488a317c27e6ac184d3b7f3ab288ac0d87d))
+
+## [0.1.6](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.1.5...@synerise/ds-slider@0.1.6) (2019-11-09)
+
+**Note:** Version bump only for package @synerise/ds-slider
+
 ## [0.1.5](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.1.4...@synerise/ds-slider@0.1.5) (2019-11-06)
 
 **Note:** Version bump only for package @synerise/ds-slider